added card health to slider in UI

+++ Assets/_TSC/_Scripts/Items/Cards/CardObject.cs
@@ -20,9 +20,9 @@ public class CardObject : ScriptableObject
     public string Name;
     public int Level = 1;
 
-    public int MaxHealth;
-    public int Health;
-    public int Attack;
+    public float MaxHealth;
+    public float Health;
+    public float Attack;
 
     public Sprite CardArtwork;
